+++ Sources/ParseSwift/API/BatchUtils.swift
@@ -23,7 +23,7 @@ typealias ParseObjectBatchResponseEncodablePointer<U> = [(Result<PointerType, Pa
 typealias RESTBatchCommandTypeEncodablePointer<T> = API.NonParseBodyCommand<ParseObjectBatchCommandEncodablePointer<T>, ParseObjectBatchResponseEncodablePointer<Encodable>> where T: Encodable
  // swiftlint:enable line_length
 
-internal struct BatchCommand<T, U>: ParseEncodable where T: ParseEncodable {
+internal struct BatchCommand<T, U>: ParseEncodable where T: ParseEncodable, U: Sendable {
     let requests: [API.Command<T, U>]
     var transaction: Bool
 }